+++ tests/smart_contracts/test_views.py
@@ -32,14 +32,14 @@ pytestmark = pytest.mark.django_db
 test_contracts_dir = Path(__file__).parent / "files" / "test_contracts"
 
 
-class TestAgreementListView:
+class TestProjectAgreementListView:
     def test_cannot_list_for_personal_project(self, app: DjangoTestApp) -> None:
         user = UserFactory()
         project = ProjectFactory(user=user)
         app.set_user(user)
 
         response = app.get(
-            reverse("agreement-list", args=[project.pk]), expect_errors=True
+            reverse("project-agreement-list", args=[project.pk]), expect_errors=True
         )
         assert response.status_code == 403
 
@@ -49,7 +49,7 @@ class TestAgreementListView:
         app.set_user(user)
 
         response = app.get(
-            reverse("agreement-list", args=[project.pk]), expect_errors=True
+            reverse("project-agreement-list", args=[project.pk]), expect_errors=True
         )
         assert response.status_code == 403
 
@@ -61,7 +61,7 @@ class TestAgreementListView:
         app.set_user(user)
         AgreementFactory.create_batch(2, project=project, assignee=organization, assigner=OrganizationFactory())
 
-        response = app.get(reverse("agreement-list", args=[project.pk]))
+        response = app.get(reverse("project-agreement-list", args=[project.pk]))
 
         assert response.status_code == 200
         assert project.agreements.count() == 2
@@ -75,7 +75,7 @@ class TestAgreementListView:
         project = ProjectFactory(organization=organization, datasets=[dataset])
         app.set_user(user)
 
-        response = app.get(reverse("agreement-list", args=[project.pk]))
+        response = app.get(reverse("project-agreement-list", args=[project.pk]))
 
         assert response.status_code == 200
         assert response.context["agreements"].count() == 0
@@ -88,14 +88,14 @@ class TestAgreementListView:
         AgreementFactory(project=project, assigner=organization)
         AgreementFactory(project=project, assigner=OrganizationFactory())
 
-        response = app.get(reverse("agreement-list", args=[project.pk]))
+        response = app.get(reverse("project-agreement-list", args=[project.pk]))
 
         assert response.status_code == 200
         assert project.agreements.count() == 2
         assert response.context["agreements"].count() == 1
 
 
-class TestAgreementDetailView:
+class TestProjectAgreementDetailView:
     def test_cannot_show_details_without_permission(
         self, app: DjangoTestApp, organization: Organization
     ) -> None:
@@ -105,7 +105,7 @@ class TestAgreementDetailView:
         agreement = AgreementFactory(project=project, assigner=organization)
 
         response = app.get(
-            reverse("agreement-detail", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-detail", args=[project.pk, agreement.pk]),
             expect_errors=True,
         )
         assert response.status_code == 403
@@ -116,7 +116,7 @@ class TestAgreementDetailView:
         project = ProjectFactory()
 
         response = app.get(
-            reverse("agreement-detail", args=[project.pk, uuid4()]), expect_errors=True
+            reverse("project-agreement-detail", args=[project.pk, uuid4()]), expect_errors=True
         )
         assert response.status_code == 404
 
@@ -130,7 +130,7 @@ class TestAgreementDetailView:
         different_project = ProjectFactory()
 
         response = app.get(
-            reverse("agreement-detail", args=[different_project.pk, agreement.pk]),
+            reverse("project-agreement-detail", args=[different_project.pk, agreement.pk]),
             expect_errors=True,
         )
         assert response.status_code == 404
@@ -144,7 +144,7 @@ class TestAgreementDetailView:
         project = ProjectFactory(organization=organization, datasets=[dataset])
         agreement = AgreementFactory(project=project, assigner=organization)
 
-        response = app.get(reverse("agreement-detail", args=[project.pk, agreement.pk]))
+        response = app.get(reverse("project-agreement-detail", args=[project.pk, agreement.pk]))
         assert response.status_code == 200
         assert response.context["agreement"] == agreement
 
@@ -159,7 +159,7 @@ class TestAgreementCreateView:
         app.set_user(user)
 
         response = app.get(
-            reverse("agreement-create", args=[project.pk]), expect_errors=True
+            reverse("project-agreement-create", args=[project.pk]), expect_errors=True
         )
         assert response.status_code == 403
 
@@ -172,7 +172,7 @@ class TestAgreementCreateView:
         app.set_user(user)
 
         response = app.get(
-            reverse("agreement-create", args=[project.pk]), expect_errors=True
+            reverse("project-agreement-create", args=[project.pk]), expect_errors=True
         )
         assert response.status_code == 404
 
@@ -185,9 +185,9 @@ class TestAgreementCreateView:
         AgreementFactory(project=project, assigner=organization)
         app.set_user(user)
 
-        response = app.get(reverse("agreement-create", args=[project.pk]))
+        response = app.get(reverse("project-agreement-create", args=[project.pk]))
         assert response.status_code == 302
-        assert response.url == reverse("agreement-list", args=[project.pk])
+        assert response.url == reverse("project-agreement-list", args=[project.pk])
 
     def test_creates_agreement_and_scopes(
         self, app: DjangoTestApp, organization: Organization, dataset: Dataset
@@ -197,13 +197,13 @@ class TestAgreementCreateView:
         project = ProjectFactory(datasets=[dataset], organization=organization)
         app.set_user(user)
 
-        response = app.get(reverse("agreement-create", args=[project.pk]))
-        form = response.forms["agreement-create"]
+        response = app.get(reverse("project-agreement-create", args=[project.pk]))
+        form = response.forms["project-agreement-create"]
         form["form-0-scopes"] = ["uapi:/test/dataset/:getall"]
         response = form.submit()
 
         assert response.status_code == 302
-        assert response.url == reverse("agreement-list", args=[project.pk])
+        assert response.url == reverse("project-agreement-list", args=[project.pk])
 
         agreement = Agreement.objects.get(project=project, assigner=organization)
         assert agreement.status == AgreementStatuses.CREATED
@@ -245,8 +245,8 @@ class TestAgreementCreateView:
         project = ProjectFactory(datasets=[dataset1, dataset2, diff_dataset], organization=organization)
         app.set_user(user)
 
-        response = app.get(reverse("agreement-create", args=[project.pk]))
-        form = response.forms["agreement-create"]
+        response = app.get(reverse("project-agreement-create", args=[project.pk]))
+        form = response.forms["project-agreement-create"]
         form["form-0-scopes"] = [
             "uapi:/test/dataset1/:getall",
             "uapi:/test/dataset2/:search",
@@ -256,7 +256,7 @@ class TestAgreementCreateView:
         response = form.submit()
 
         assert response.status_code == 302
-        assert response.url == reverse("agreement-list", args=[project.pk])
+        assert response.url == reverse("project-agreement-list", args=[project.pk])
 
         assert Agreement.objects.filter(project=project).count() == 2
         assert set(
@@ -287,8 +287,8 @@ class TestAgreementCreateView:
         project = ProjectFactory(datasets=[dataset, dataset2], organization=organization)
         AgreementFactory(project=project, assigner=organization)
 
-        response = app.get(reverse("agreement-create", args=[project.pk]))
-        form = response.forms["agreement-create"]
+        response = app.get(reverse("project-agreement-create", args=[project.pk]))
+        form = response.forms["project-agreement-create"]
 
         assert form.fields.get("form-0-scopes")
         assert not form.fields.get("form-1-scopes")
@@ -309,7 +309,7 @@ class TestAgreementCreateView:
             "form-0-id": organization.id,
             "form-0-scopes": ["bad_scope"],
         }
-        response = app.post(reverse("agreement-create", args=[project.pk]), data)
+        response = app.post(reverse("project-agreement-create", args=[project.pk]), data)
 
         assert response.status_code == 200
         assert Agreement.objects.filter(project=project).count() == 0
@@ -321,7 +321,7 @@ class TestAgreementCreateView:
         project = ProjectFactory(user=user, datasets=[dataset])
         app.set_user(user)
 
-        response = app.get(reverse("agreement-create", args=[project.pk]), status=403)
+        response = app.get(reverse("project-agreement-create", args=[project.pk]), status=403)
 
         assert response.status_code == 403
 
@@ -361,7 +361,7 @@ class TestAgreementSubmit:
 
         # Act
         response = app.post(
-            reverse("agreement-submit", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-submit", args=[project.pk, agreement.pk]),
             {
                 "assignee_representative": contact.pk
             }
@@ -406,7 +406,7 @@ class TestAgreementSubmit:
 
         # Act
         response = app.post(
-            reverse("agreement-submit", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-submit", args=[project.pk, agreement.pk]),
             {
                 "assignee_representative": contact.pk,
             },
@@ -453,7 +453,7 @@ class TestAgreementSubmit:
 
         # Act
         response = app.post(
-            reverse("agreement-submit", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-submit", args=[project.pk, agreement.pk]),
             {
                 "assignee_representative": contact.pk
             },
@@ -500,7 +500,7 @@ class TestAgreementSubmit:
 
         # Act
         response = app.post(
-            reverse("agreement-submit", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-submit", args=[project.pk, agreement.pk]),
             {
                 "assignee_representative": contact.pk
             },
@@ -561,7 +561,7 @@ class TestAgreementSubmit:
 
         # Act
         response = app.post(
-            reverse("agreement-submit", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-submit", args=[project.pk, agreement.pk]),
             {
                 "assignee_representative": contact.pk
             }
@@ -634,7 +634,7 @@ class TestAgreementApprove:
 
         # Act
         response = app.post(
-            reverse("agreement-approve", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-approve", args=[project.pk, agreement.pk]),
             {
                 "template": template.pk,
                 "assigner_representative": assigner_contact.pk,
@@ -709,7 +709,7 @@ class TestAgreementApprove:
 
         # Act
         response = app.post(
-            reverse("agreement-approve", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-approve", args=[project.pk, agreement.pk]),
             {
                 "template": template.pk,
                 "assigner_representative": assigner_contact.pk,
@@ -785,7 +785,7 @@ class TestAgreementApprove:
 
         # Act
         response = app.post(
-            reverse("agreement-approve", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-approve", args=[project.pk, agreement.pk]),
             {
                 "template": template.pk,
                 "assigner_representative": assigner_contact.pk,
@@ -861,7 +861,7 @@ class TestAgreementApprove:
 
         # Act
         response = app.post(
-            reverse("agreement-approve", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-approve", args=[project.pk, agreement.pk]),
             {
                 "template": template.pk,
                 "assigner_representative": assigner_contact.pk,
@@ -951,7 +951,7 @@ class TestAgreementApprove:
 
         # Act
         response = app.post(
-            reverse("agreement-approve", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-approve", args=[project.pk, agreement.pk]),
             {
                 "template": template.pk,
                 "assigner_representative": assigner_contact.pk,
@@ -1033,7 +1033,7 @@ class TestAgreementForm:
 
         # Act
         response = app.post(
-            reverse("agreement-form", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-form", args=[project.pk, agreement.pk]),
             {}
         )
 
@@ -1190,7 +1190,7 @@ class TestAgreementForm:
 
         # Act
         response = app.post(
-            reverse("agreement-form", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-form", args=[project.pk, agreement.pk]),
             {},
             expect_errors=True,
         )
@@ -1265,7 +1265,7 @@ class TestAgreementForm:
 
         # Act
         response = app.post(
-            reverse("agreement-form", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-form", args=[project.pk, agreement.pk]),
             {},
             expect_errors=True,
         )
@@ -1340,7 +1340,7 @@ class TestAgreementForm:
 
         # Act
         response = app.post(
-            reverse("agreement-form", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-form", args=[project.pk, agreement.pk]),
             {},
             expect_errors=True,
         )
@@ -1422,7 +1422,7 @@ class TestAgreementForm:
 
         # Act
         response = app.post(
-            reverse("agreement-form", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-form", args=[project.pk, agreement.pk]),
             {},
             expect_errors=True,
         )
@@ -1466,7 +1466,7 @@ class TestAgreementInitiate:
 
         # Act
         response = app.post(
-            reverse("agreement-initiate", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-initiate", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",
@@ -1513,7 +1513,7 @@ class TestAgreementInitiate:
 
         # Act
         response = app.post(
-            reverse("agreement-initiate", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-initiate", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",
@@ -1570,7 +1570,7 @@ class TestAgreementInitiate:
 
         # Act
         response = app.post(
-            reverse("agreement-initiate", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-initiate", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",
@@ -1623,7 +1623,7 @@ class TestAgreementInitiate:
 
         # Act
         response = app.post(
-            reverse("agreement-initiate", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-initiate", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",
@@ -1710,7 +1710,7 @@ class TestAgreementSign:
 
         # Act
         response = app.post(
-            reverse("agreement-sign", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-sign", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",
@@ -1805,7 +1805,7 @@ class TestAgreementSign:
 
         # Act
         response = app.post(
-            reverse("agreement-sign", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-sign", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",
@@ -1893,7 +1893,7 @@ class TestAgreementSign:
 
         # Act
         response = app.post(
-            reverse("agreement-sign", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-sign", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",
@@ -1979,7 +1979,7 @@ class TestAgreementSign:
 
         # Act
         response = app.post(
-            reverse("agreement-sign", args=[project.pk, agreement.pk]),
+            reverse("project-agreement-sign", args=[project.pk, agreement.pk]),
             upload_files=[
                 (
                     "file",